shadcn integration, config update, path alias update

+++ eslint.config.js
@@ -30,7 +30,7 @@ export default defineConfig([
         "warn",
         { allowConstantExport: true },
       ],
-      "quotes": [2, "single", { "avoidEscape": true }],
+      "quotes": [2, "double", { "avoidEscape": true }],
       "semi": [2, "always"],
     },
   },
